- Separates the campaign button into two components
